% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate.SField}
\alias{aggregate}
\alias{aggregate.SField}
\alias{aggregate.SLattice}
\alias{aggregate.SObjects}
\title{aggregate SField, SObjects or SLattice objects}
\usage{
aggregate(x, ...)

\method{aggregate}{SObjects}(x, by, FUN = mean, ...)

\method{aggregate}{SLattice}(x, by, FUN = mean, ...)
}
\arguments{
\item{x}{object of class \link{SField}, \link{SObjects}, or \link{SLattice}}

\item{by}{aggregation predicate}

\item{FUN}{aggregation function}

\item{...}{arguments passed on to \code{FUN}}
}
\description{
aggregation methods for objects of class SField, SObjects or SLattice
}

