% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field-methods.R
\name{full.coverage}
\alias{full.coverage}
\title{assess whether function has complete coverage over the domain}
\usage{
full.coverage(x)
}
\arguments{
\item{x}{object of class \link{SField-class}}
}
\value{
logical; TRUE if support is "point" and area size of the observations is identical to that of the domain; this implies we have a coverage with knowledge of all points in the domain.
}
\description{
fields with point support, defined for all areas in the domain are completely known; this function verifies that this is the case
}
\note{
what the function does is not sufficient: identical area size does not guarantee that areas are identical
}

