% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/field-methods.R
\name{getArea}
\alias{getArea}
\title{compute area of a Spatial* object}
\usage{
getArea(x)
}
\arguments{
\item{x}{object of a subclass of \link[sp]{Spatial}}
}
\value{
the area of the features (grid cells or polygons) in the object
}
\description{
compute the area of an object of one of the subclasses of \link[sp]{Spatial}
}
\details{
area of gridded or polygon objects deriving from \link[sp]{Spatial}, zero
for other classes (points, lines)
}
\examples{
library(sp)
demo(meuse, ask = FALSE, echo = FALSE)
getArea(meuse)
getArea(meuse.area)
}

