% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-field.R
\docType{methods}
\name{initialize,SField-method}
\alias{initialize,SField-method}
\title{SField initialize function}
\usage{
\S4method{initialize}{SField}(.Object, observations, domain,
  cellsArePoints = NA)
}
\arguments{
\item{.Object}{(ignore)}

\item{observations}{object of one of the sublasses of \link[sp]{Spatial}}

\item{domain}{object of class \link{SExtent}, or subclass of \link[sp]{Spatial}}

\item{cellsArePoints}{logical; are grid cells to be taken as point (support) values?}
}
\value{
object of class \link{SField-class}
}
\description{
initializes (creates) SField objects
}
\seealso{
\link{SObjects}
}

