% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-entity.R
\docType{methods}
\name{initialize,SObjects-method}
\alias{initialize,SObjects-method}
\title{SObjects initialize function}
\usage{
\S4method{initialize}{SObjects}(.Object, observations, window)
}
\arguments{
\item{.Object}{object to initialize}

\item{observations}{object of one of the sublasses of \link{Spatial}}

\item{window}{object of class \link{SExtent}}
}
\value{
object of class \link{SObjects-class}
}
\description{
initializes (creates) SObjects objects
}
\note{
this function is not called directly
}
\seealso{
\link{SField}
}

