% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class-entity.R
\docType{methods}
\name{initialize,SpatialEntities-method}
\alias{initialize,SpatialEntities-method}
\title{SpatialEntities initialize function}
\usage{
\S4method{initialize}{SpatialEntities}(.Object, observations, window)
}
\arguments{
\item{.Object}{object to initialize}

\item{observations}{object of one of the sublasses of \link{Spatial}}

\item{window}{object of class \link{Window}}
}
\value{
object of class \link{SpatialEntities-class}
}
\description{
initializes (creates) SpatialEntities objects
}
\note{
this function is not called directly
}
\seealso{
\link{SpatialField}
}

