% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class-field.R
\docType{methods}
\name{initialize,SpatialField-method}
\alias{initialize,SpatialField-method}
\title{SpatialField initialize function}
\usage{
\S4method{initialize}{SpatialField}(.Object, observations, domain,
  cellsArePoints = NA)
}
\arguments{
\item{.Object}{(ignore)}

\item{observations}{object of one of the sublasses of \link[sp]{Spatial}}

\item{domain}{object of class \link{Window}, or subclass of \link[sp]{Spatial}}

\item{cellsArePoints}{logical; are grid cells to be taken as point (support) values?}
}
\value{
object of class \link{SpatialField-class}
}
\description{
initializes (creates) SpatialField objects
}
\seealso{
\link{SpatialEntities}
}

