% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/class-window.R
\docType{methods}
\name{initialize,Window-method}
\alias{initialize,Window-method}
\title{initializes (creates) Window objects}
\usage{
\S4method{initialize}{Window}(.Object, sp)
}
\arguments{
\item{.Object}{(ignore)}

\item{sp}{object of one of the sublasses of \link[sp]{Spatial}}
}
\value{
object of class \link{Window-class}
}
\description{
initializes (creates) Window objects
}
\note{
this function should not be called directly, use \link{Window} instead
}

