% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/entities-methods.R, R/field-methods.R
\docType{methods}
\name{interpolate,formula,SpatialEntities,ANY-method}
\alias{interpolate}
\alias{interpolate,formula,SpatialAggregation,SpatialAggregation-method}
\alias{interpolate,formula,SpatialAggregation,SpatialField-method}
\alias{interpolate,formula,SpatialEntities,ANY-method}
\alias{interpolate,formula,SpatialField,SpatialAggregation-method}
\alias{interpolate,formula,SpatialField,SpatialField-method}
\alias{interpolate,formula,SpatialField,missing-method}
\title{Method interpolate}
\usage{
\S4method{interpolate}{formula,SpatialEntities,ANY}(formula, data, newdata, ...)

\S4method{interpolate}{formula,SpatialField,SpatialField}(formula, data,
  newdata, ...)

\S4method{interpolate}{formula,SpatialField,missing}(formula, data, ...,
  ncells = 5000)

\S4method{interpolate}{formula,SpatialField,SpatialAggregation}(formula, data,
  newdata, ..., ncells = 5000)

\S4method{interpolate}{formula,SpatialAggregation,SpatialField}(formula, data,
  newdata, model, ...)

\S4method{interpolate}{formula,SpatialAggregation,SpatialAggregation}(formula,
  data, newdata, model, ...)
}
\arguments{
\item{formula}{specifing which variable should be interpolated, see also \link[gstat]{krige}}

\item{data}{object of class \link{SpatialField-class}}

\item{newdata}{target locations; if missing, points are chosen from the domain of \code{data}}

\item{...}{passed on to \link[gstat]{krige}}

\item{ncells}{in case no newdata is provided and point support interpolations are computed over a polygon area, the approximate number of grid cells (default 5000)}

\item{model}{covariance model, see \link[gstat]{krige} and \link[gstat]{vgm}}
}
\value{
object of class \link{SpatialField-class}
}
\description{
Method interpolate

interpolate SpatialField data
}
\examples{
library(sp)
demo(meuse, ask = FALSE, echo = FALSE)
sf = SpatialField(meuse, Window(meuse.grid))
p = interpolate(zinc~1, sf)
spplot(p)
sf = SpatialField(meuse, Window(meuse.area))
p = interpolate(zinc~1, sf)
spplot(p)
}
\seealso{
\link[gstat]{krige}
}

