% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/class-field.R
\name{is.complete}
\alias{is.complete}
\title{assess whether function has complete coverage over the domain}
\usage{
is.complete(x)
}
\arguments{
\item{x}{object of class \link{SpatialField-class}}
}
\value{
logical; TRUE if support is "point" and area size of the observations is identical to that of the domain; this implies we have a coverage with knowledge of all points in the domain.
}
\description{
fields with point support, defined for all areas in the domain are completely known; this function verifies that this is the case
}
\note{
what the function does is not sufficient: identical area size guaranties not that areas are identical
}

