% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{is.sum}
\alias{is.sum}
\title{check whether x equals function sum}
\usage{
is.sum(x)
}
\arguments{
\item{x}{the function to check}
}
\description{
this function checks whether its argument equals function sum
}
\examples{
is.sum(sum)
is.sum(sum(1:3))
is.sum(mean)
is.sum(function(x) sum(x))
}

