% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/mss.R
\name{mss_error}
\alias{mss_error}
\title{assign whether non-meaningful actions should result in an error
assign whether non-meaningful actions should result in an error}
\usage{
mss_error(error = TRUE)
}
\arguments{
\item{error}{logical; should a non-meaningful activity result in an error?}
}
\value{
the logical value assigned
}
\description{
assign whether non-meaningful actions should result in an error
assign whether non-meaningful actions should result in an error
}

