% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/plot.R
\docType{methods}
\name{plot,SpatialAggregation,missing-method}
\alias{plot,SpatialAggregation,missing-method}
\alias{plot,SpatialEntities,missing-method}
\alias{plot,SpatialField,missing-method}
\title{plot method for SpatialAggregation objects}
\usage{
\S4method{plot}{SpatialAggregation,missing}(x, y, ...)

\S4method{plot}{SpatialField,missing}(x, y, ..., bg = grey(0.7))

\S4method{plot}{SpatialEntities,missing}(x, ..., bg = grey(0.7))
}
\arguments{
\item{x}{object of class \link{SpatialAggregation}, \link{SpatialField} or \link{SpatialEntities}}

\item{y}{should be missing}

\item{...}{passed on to \link[sp]{plot,Spatial,missing-method}}

\item{bg}{background colour for domain or observation window (default grey(.7))}
}
\description{
plot method for SpatialAggregation objects
}
\examples{
library(sp)
demo(meuse, ask = FALSE, echo = FALSE)
plot(SpatialAggregation(meuse.grid))
plot(SpatialAggregation(meuse.area))
plot(SpatialField(meuse, meuse.grid))
plot(SpatialField(meuse, meuse.area))
plot(SpatialEntities(meuse, meuse.grid))
plot(SpatialEntities(meuse, meuse.area))
}

