% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-window.R
\docType{class}
\name{SExtent}
\alias{SExtent}
\alias{SExtent-class}
\alias{SExtentOrNULL-class}
\alias{initialize,SExtent-method}
\title{The SExtent class and Constructor Function: Domain and Observed SExtent}
\usage{
SExtent(...)
}
\arguments{
\item{...}{named parameter, containing 
\describe{
  \item{\code{area}}{object of a subclass of \link[sp]{Spatial}}
}}
}
\value{
object of class \link{SExtent}
}
\description{
Class for defining domain of a field, or observed window for entities
}
\details{
In practice, spatial fields are defined for a limited domain. Spatial
entities can form an exhaustive set for an area called the observed
window. For both, the SExtent class is used to specify this area.
}
\note{
if \code{area} is not of (a subclass of) class \link[sp]{SpatialPolygons}, \link[sp]{SpatialPixels}, or \link[sp]{SpatialGrid}, an error results.
}
\section{Slots}{
 
 \describe{
   \item{\code{area}}{object of a subclass of \link[sp]{Spatial}}
 }
}
\author{
Edzer Pebesma
}

