% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-field.R
\docType{class}
\name{SField-class}
\alias{$,SField-method}
\alias{SField}
\alias{SField-class}
\alias{[,SField-method}
\alias{[[,SField,ANY,missing-method}
\alias{[[<-,SField,ANY,missing-method}
\alias{over,SField,SField-method}
\alias{over,SField,SLattice-method}
\alias{spplot,SField-method}
\title{The SField class and constructor function}
\usage{
SField(observations, domain, cellsArePoints = NA)
}
\arguments{
\item{observations}{object of one of the sublasses of \link[sp]{Spatial}}

\item{domain}{object of class \link{Window}, or of a subclass of \link[sp]{Spatial}; if missing, it is assumed to be identical to \code{observations}}

\item{cellsArePoints}{logical; do grid cell values reflect point values at the grid cell centre (TRUE) or constant point values throughout the whole grid cell (FALSE)?}
}
\value{
object of class \link{SField-class}
}
\description{
SField: a Class for Spatial Fields (Objects)
}
\details{
A class to store spatial fields, such as temperatures, elevation,
land use, and so on, which have point support (footprint), along with the
domain for which the set observation are valid e.g. to make interpolations.

Fields are representations of continuous phenomena: at every point in continuous
space, a value exists. Observations on fields are, by necessity, discrete and
countable when they refer to points, but may also consist of areas in which case
an infinite number of points with constant value in the area is represented.

SField objects can be generated from points, lines, polygons or grids. For
all classes except points, the attribute values are assumed to be constant and
identical for all points along the line (lines), or over the area (polygons, grids).

The domain of a SField object is the area for which the observations are
considered relevant, e.g. by affording interpolation or aggregation.
}
\note{
If no domain is supplied, the domain is set to the collection of features. A warning is issued if one or more of the features are (completely) outside the domain.
}
\section{Slots}{
 
 \describe{
   \item{\code{observations}:}{object of a subclass of \link[sp]{Spatial}}
   \item{\code{domain}:}{object of class \link{WindowOrNULL-class}}
   \item{\code{cellsArePoints}}{logical; do grid cell values reflect point values at the grid cell centre (TRUE) or constant point values throughout the whole grid cell (FALSE)?}
 }
}
\examples{
library(sp)
demo(meuse, ask = FALSE, echo = FALSE)
m = SField(meuse, meuse.area)
}
\author{
Edzer Pebesma
}
\seealso{
\link{SObjects}
}

