% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-entity.R
\docType{class}
\name{SObjects-class}
\alias{$,SObjects-method}
\alias{SObjects}
\alias{SObjects-class}
\alias{[,SObjects-method}
\alias{[[,SObjects,ANY,missing-method}
\alias{[[<-,SObjects,ANY,missing-method}
\title{The SObjects class and constructor function}
\usage{
SObjects(observations, window)
}
\arguments{
\item{observations}{object of one of the sublasses of \link{Spatial}}

\item{window}{object of class \link{SExtent} or \link{Spatial}}
}
\value{
object of class \link{SObjects-class}
}
\description{
SObjects: a Class for Spatial Entities (Objects)
}
\details{
A class to store spatial entities, such as trees, buildings and so on,
along with the observed window, the area for which the set of
entities is exhaustive. It is assumed that for all points in the 
observed window where there are no objects, the space is known to be empty.
}
\note{
If no window is supplied, the window is set to the collection of features, and a warning is issued. Support is assumed to be that of the features (points, polygons, lines, grid cells) in the \code{observations} object.
}
\section{Slots}{
 
 \describe{
   \item{\code{observations}}{object of a subclass of \link[sp]{Spatial}}
   \item{\code{window}}{object of class \link{SExtentOrNULL-class}}
 }
}
\examples{
library(sp)
demo(meuse, ask = FALSE, echo = FALSE)
m = SObjects(meuse, meuse.area)
}
\author{
Edzer Pebesma
}
\seealso{
\link{SField}
}

