% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/class-field.R
\docType{class}
\name{SpatialField-class}
\alias{$,SpatialField-method}
\alias{SpatialField}
\alias{SpatialField-class}
\alias{[,SpatialField-method}
\alias{spplot,SpatialField-method}
\title{The SpatialField class and constructor function}
\usage{
SpatialField(...)
}
\arguments{
\item{...}{named arguments, according to the following list:
 \describe{
   \item{sp}{object of one of the sublasses of \link{Spatial}}
   \item{support}{object of class \link{Support}, or one of \code{"point"}, \code{"feature"}}
   \item{domain}{object of class \link{Window}, or of a subclass of \link[sp]{Spatial} (typically: \link[sp]{SpatialPolygons}, \link[sp]{SpatialPixels} or \link[sp]{SpatialGrid})}
   \item{sp_equals_domain}{logical; indicates whether the data of
   \code{sp} is identical to the domain}
 }}
}
\value{
object of class \link{SpatialField-class}
}
\description{
SpatialField: a Class for Spatial Fields (Objects)
}
\details{
A class to store spatial fields, such as temperatures, elevation,
land use, and so on, along with their support (or footprint) and
domain for which the set observation are valid and interpolation makes sense.
}
\note{
If no domain is supplied, the domain is set to the collection of features. If no support is specified, the support is set to that of the features (points, polygons, lines, grid cells) in the \code{sp} object, and a warning is issued. A warning is issued if one or more of the features are (completely) outside the domain.
}
\section{Slots}{

 \describe{
   \item{\code{sp}}{object of a subclass of \link[sp]{Spatial}}
   \item{\code{support}}{object of class \link{Support}}
   \item{\code{domain}}{object of class \link{Window}}
   \item{\code{sp_equals_domain}}{logical; is \code{sp} identical to \code{domain}?}
 }
}
\examples{
library(sp)
demo(meuse, ask = FALSE, echo = FALSE)
m = SpatialEntities(meuse, "point", meuse.area)
}
\author{
Edzer Pebesma
}
\seealso{
\link{SpatialEntities}
}

