% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/class-support.R
\docType{class}
\name{Support}
\alias{Support}
\alias{Support-class}
\title{The Support class and constructor function: spatial support}
\usage{
Support(what, fn)
}
\arguments{
\item{what}{character: one of: \code{"point"}, \code{"feature"}, \code{"function"}; if omitted, \code{"features"} is chosen an a warning is given}

\item{fn}{function computing the support}
}
\value{
object of class \link{Support}
}
\description{
Class for defining support of observations or predictions
}
\details{
(Spatial) Support is the physical size (length, area, volume) corresponding
to measurements, or predictions. This class informs objects of the subclasses
of \link[sp]{Spatial} about their support.
}
\note{
a function computing the support should provide as input value the feature (point, line, or polygon) for which a support is to be computed.
}
\section{Slots}{

 \describe{
   \item{\code{what}}{character; one of: \code{"point"}, \code{"feature"}, \code{"function"}}
   \item{\code{fn}}{function, (optionally) computing support}
 }
}
\author{
Edzer Pebesma
}

