% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class-window.R
\docType{class}
\name{Window}
\alias{Window}
\alias{Window-class}
\alias{WindowOrNULL-class}
\alias{initialize,Window-method}
\title{The Window class and Constructor Function: Domain and Observed Window}
\usage{
Window(...)
}
\arguments{
\item{...}{named parameter, containing
\describe{
  \item{\code{area}}{object of a subclass of \link[sp]{Spatial}}
}}
}
\value{
object of class \link{Window}
}
\description{
Class for defining domain of a field, or observed window for entities
}
\details{
In practice, spatial fields are defined for a limited domain. Spatial
entities can form an exhaustive set for an area called the observed
window. For both, the Window class is used to specify this area.
}
\note{
if \code{area} is not of (a subclass of) class \link[sp]{SpatialPolygons}, \link[sp]{SpatialPixels}, or \link[sp]{SpatialGrid}, an error results.
}
\section{Slots}{

 \describe{
   \item{\code{area}}{object of a subclass of \link[sp]{Spatial}}
 }
}
\author{
Edzer Pebesma
}

