% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate.SpatialField}
\alias{aggregate}
\alias{aggregate.SpatialEntities}
\alias{aggregate.SpatialField}
\title{aggregate SpatialField or SpatialEntities objects}
\usage{
aggregate(x, ...)

\method{aggregate}{SpatialEntities}(x, by, FUN = mean, ...)
}
\arguments{
\item{x}{object of class \code{SpatialField}}

\item{by}{aggregation predicate}

\item{FUN}{aggregation function}

\item{...}{arguments passed on to \code{FUN}}
}
\description{
aggregation methods for objects of class SpatialField or SpatialEntities
}

