% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/entities-methods.R
\name{density.SpatialEntities}
\alias{density}
\alias{density.SpatialEntities}
\title{SpatialEntities density method}
\usage{
\method{density}{SpatialEntities}(x, bandwidth, newdata, ncells = 5000, ...)
}
\arguments{
\item{x}{object of class \link{SpatialEntities-class}}

\item{bandwidth}{bandwidth parameter (see \link[MASS]{kde2d})}

\item{newdata}{target grid; if omitted, a grid over the window is created}

\item{ncells}{in case no newdata is provided and window is a polygon, the approximate number of grid cells for the grid created}

\item{...}{ignored}
}
\value{
object of class \link{SpatialField-class}
}
\description{
density estimate for SpatialEntities data
}

