% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/class-field.R
\docType{methods}
\name{initialize,SpatialField-method}
\alias{initialize,SpatialField-method}
\title{SpatialField initialize function}
\usage{
\S4method{initialize}{SpatialField}(.Object, sp, support, domain)
}
\arguments{
\item{.Object}{(ignore)}

\item{sp}{object of one of the sublasses of \link[sp]{Spatial}}

\item{support}{object of class \link{Support}, or one of "point" and "feature"}

\item{domain}{object of class \link{Window}, or subclass of \link[sp]{Spatial}}
}
\value{
object of class \link{SpatialField-class}
}
\description{
initializes (creates) SpatialField objects
}
\seealso{
\link{SpatialEntities}
}

