% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/field-methods.R
\docType{methods}
\name{interpolate}
\alias{interpolate}
\alias{interpolate,formula,SpatialField,SpatialField-method}
\title{Method interpolate}
\usage{
\S4method{interpolate}{formula,SpatialField,SpatialField}(formula, data,
  newdata, ...)
}
\arguments{
\item{formula}{specifing which variable should be interpolated, see also \link[gstat]{krige}}

\item{data}{object of class \link{SpatialField-class}}

\item{newdata}{target locations; if missing, points are chosen from the domain of \code{data}}

\item{...}{passed on to \link[gstat]{krige}}

\item{ncells}{in case no newdata is provided and point support interpolations are computed over a polygon area, the approximate number of grid cells (default 5000)}
}
\value{
object of class \link{SpatialField-class}
}
\description{
Method interpolate

interpolate SpatialField data
}
\seealso{
\link[gstat]{krige}
}

