% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field-methods.R, R/entities-methods.R
\docType{methods}
\name{interpolate}
\alias{interpolate}
\alias{interpolate,formula,SField,SField-method}
\alias{interpolate,formula,SField,SLattice-method}
\alias{interpolate,formula,SField,missing-method}
\alias{interpolate,formula,SLattice,SField-method}
\alias{interpolate,formula,SLattice,SLattice-method}
\alias{interpolate,formula,SObjects,ANY-method}
\title{Method interpolate}
\usage{
\S4method{interpolate}{formula,SField,SField}(formula, data, newdata, ...)

\S4method{interpolate}{formula,SField,missing}(formula, data, ...,
  ncells = 5000)

\S4method{interpolate}{formula,SField,SLattice}(formula, data, newdata, ...,
  ncells = 5000)

\S4method{interpolate}{formula,SLattice,SField}(formula, data, newdata, model,
  ...)

\S4method{interpolate}{formula,SLattice,SLattice}(formula, data, newdata, model,
  ...)

\S4method{interpolate}{formula,SObjects,ANY}(formula, data, newdata, ...)
}
\arguments{
\item{formula}{specifing which variable should be interpolated, see also \link[gstat]{krige}}

\item{data}{object of class \link{SField-class}}

\item{newdata}{target locations; if missing, points are chosen from the domain of \code{data}}

\item{...}{passed on to \link[gstat]{krige}}

\item{ncells}{in case no newdata is provided and point support interpolations are computed over a polygon area, the approximate number of grid cells (default 5000)}

\item{model}{covariance model, see \link[gstat]{krige} and \link[gstat]{vgm}}
}
\value{
object of class \link{SField-class}
}
\description{
Method interpolate

interpolate SField data
}
\examples{
library(sp)
demo(meuse, ask = FALSE, echo = FALSE)
sf = SField(meuse, Window(meuse.grid))
p = interpolate(zinc~1, sf)
spplot(p)
sf = SField(meuse, Window(meuse.area))
p = interpolate(zinc~1, sf)
spplot(p)
}
\seealso{
\link[gstat]{krige}
}

