% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\docType{methods}
\name{plot,SLattice,missing-method}
\alias{plot,SField,missing-method}
\alias{plot,SLattice,missing-method}
\alias{plot,SObjects,missing-method}
\title{plot method for SLattice objects}
\usage{
\S4method{plot}{SLattice,missing}(x, y, ...)

\S4method{plot}{SField,missing}(x, y, ..., bg = grey(0.7))

\S4method{plot}{SObjects,missing}(x, ..., bg = grey(0.7))
}
\arguments{
\item{x}{object of class \link{SLattice}, \link{SField} or \link{SObjects}}

\item{y}{should be missing}

\item{...}{passed on to \link[sp]{plot,Spatial,missing-method}}

\item{bg}{background colour for domain or observation window (default grey(.7))}
}
\description{
plot method for SLattice objects
}
\examples{
library(sp)
demo(meuse, ask = FALSE, echo = FALSE)
plot(SLattice(meuse.grid))
plot(SLattice(meuse.area))
plot(SField(meuse, meuse.grid))
plot(SField(meuse, meuse.area))
plot(SObjects(meuse, meuse.grid))
plot(SObjects(meuse, meuse.area))
}

