% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/class-entity.R
\docType{class}
\name{SpatialEntities-class}
\alias{SpatialEntities}
\alias{SpatialEntities-class}
\title{The SpatialEntities class and constructor function}
\usage{
SpatialEntities(...)
}
\arguments{
\item{...}{named arguments from the following list:
\describe{
  \item{\code{observations}}{object of one of the sublasses of \link{Spatial}}
  \item{\code{window}}{object of class \link{Window} or \link{Spatial}}
}}
}
\value{
object of class \link{SpatialEntities-class}
}
\description{
SpatialEntities: a Class for Spatial Entities (Objects)
}
\details{
A class to store spatial entities, such as trees, buildings and so on,
along with the observed window, the area for which the set of
entities is exhaustive
}
\note{
If no window is supplied, the window is set to the collection of features, and a warning is issued. Support is assumed to be that of the features (points, polygons, lines, grid cells) in the \code{observations} object.
}
\section{Slots}{

 \describe{
   \item{\code{observations}}{object of a subclass of \link[sp]{Spatial}}
   \item{\code{window}}{object of class \link{Window}}
 }
}
\examples{
library(sp)
demo(meuse, ask = FALSE, echo = FALSE)
m = SpatialEntities(meuse, meuse.area)
}
\author{
Edzer Pebesma
}
\seealso{
\link{SpatialField}
}

