% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/class-field.R
\docType{class}
\name{SpatialField-class}
\alias{$,SpatialField-method}
\alias{SpatialField}
\alias{SpatialField-class}
\alias{[,SpatialField-method}
\alias{spplot,SpatialField-method}
\title{The SpatialField class and constructor function}
\usage{
SpatialField(...)
}
\arguments{
\item{...}{named arguments, according to the following list:
 \describe{
   \item{observations}{object of one of the sublasses of \link[sp]{Spatial}}
   \item{domain}{object of class \link{Window}, or of a subclass of \link[sp]{Spatial} (typically: \link[sp]{SpatialPolygons}, \link[sp]{SpatialPixels} or \link[sp]{SpatialGrid})}
   \item{observations_equal_domain}{logical; indicates whether the
   \code{observations} are identical to the domain}
 }}
}
\value{
object of class \link{SpatialField-class}
}
\description{
SpatialField: a Class for Spatial Fields (Objects)
}
\details{
A class to store spatial fields, such as temperatures, elevation,
land use, and so on, which have point support (footprint), along with the
domain for which the set observation are valid e.g. to make interpolations.
}
\note{
If no domain is supplied, the domain is set to the collection of features. A warning is issued if one or more of the features are (completely) outside the domain.
}
\section{Slots}{

 \describe{
   \item{\code{observations}}{object of a subclass of \link[sp]{Spatial}}
   \item{\code{domain}}{object of class \link{Window}}
   \item{\code{observations_equal_domain}}{logical; are the \code{observations} identical to the \code{domain}?}
 }
}
\examples{
library(sp)
demo(meuse, ask = FALSE, echo = FALSE)
m = SpatialField(meuse, meuse.area)
}
\author{
Edzer Pebesma
}
\seealso{
\link{SpatialEntities}
}

